requirements.txt
flask
flask-sqlalchemy
flask-login
python-dotenv
werkzeug
pymysql
google-generativeai
requests